requirements.txt
transformers==4.38.2
pytorch-lightning==2.2.1
pytorch-nlp==0.5.0
scikit-learn==0.24.2
pandas==1.1.5